pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "python-jose>=3.5.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg2>=2.9.11",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
